Adds redis
