+++ server/package.json
@@ -4,12 +4,18 @@
   "description": "",
   "main": "src/index.js",
   "scripts": {
-    "start": "node src/index.js"
+    "start": "node src/index.js",
+    "postinstall": "link-module-alias"
   },
   "author": "kaihsien",
   "license": "ISC",
   "dependencies": {
     "express": "^4.17.1",
+    "link-module-alias": "^1.2.0",
+    "redis": "^3.1.2",
     "socket.io": "^4.1.2"
+  },
+  "_moduleAliases": {
+    "@": "src"
   }
 }
