+++ server/src/db/index.js
@@ -1,0 +1,11 @@
+const redis = require("./redis");
+
+const createRoom = (roomNumber, room) => {
+    redis.client
+}
+
+module.exports = {
+    db: redis,
+    setGameForRoom: redis.setGameForRoom,
+    getGameForRoom: redis.getGameForRoom,
+}
